+++ src/commerce.tsx
@@ -19,53 +19,62 @@ type CommerceProps = {
 type CommerceConfig = {
   token: string;
   domain: string;
-  currencyCode: string;
+  currencyCode?: string;
+  locale?: string;
+  sessionToken?: string;
 };
 
 type CommerceContextValue = {
-  token: string;
-  domain: string;
-  client?: ClientType;
+  client: ClientType;
   shop: Shop;
   checkout: Cart;
-  currencyCode: string;
   updateCheckout: (cart: Cart | undefined) => void;
+  currencyCode: string;
+  locale: string;
+  sessionToken: string;
 };
 
-const TOKEN = 'nextjs-commerce-shopify-token';
-
-const getCheckoutIdFromStorage = () => {
+const getCheckoutIdFromStorage = (token: string) => {
   if (window && window.sessionStorage) {
-    return window.sessionStorage.getItem(TOKEN);
+    return window.sessionStorage.getItem(token);
   }
 
   return null;
 };
 
-const setCheckoutIdInStorage = (id: string | number) => {
+const setCheckoutIdInStorage = (token: string, id: string | number) => {
   if (window && window.sessionStorage) {
-    return window.sessionStorage.setItem(TOKEN, id + '');
+    return window.sessionStorage.setItem(token, id + '');
   }
 };
 
+const defaults = {
+  locale: 'en-US',
+  sessionToken: 'nextjs-commerce-shopify-token'
+};
+
 export function CommerceProvider({ children, config }: CommerceProps) {
+  const sessionToken = config.sessionToken || defaults.sessionToken;
+  const locale = config.locale || defaults.locale;
+
   const client = Client.buildClient({
     storefrontAccessToken: config.token,
-    domain: config.domain
+    domain: config.domain,
+    language: locale
   }) as ClientType;
 
   const [shop, setShop] = useState<Shop>();
   const [checkout, setCheckout] = useState<Cart>();
 
   const fetchShopify = async () => {
-    const shopInfo = await client.shop.fetchInfo();
+    const shopInfo: Shop = await client.shop.fetchInfo();
     let checkoutResource: Cart;
 
     const checkoutOptions = {
-      presentmentCurrencyCode: config.currencyCode
+      presentmentCurrencyCode: config.currencyCode || shopInfo?.currencyCode
     };
 
-    let checkoutId = getCheckoutIdFromStorage();
+    let checkoutId = getCheckoutIdFromStorage(sessionToken);
 
     // we could have a cart id stored in session storage
     // user could be refreshing or navigating back and forth
@@ -80,7 +89,7 @@ export function CommerceProvider({ children, config }: CommerceProps) {
       checkoutResource = await client.checkout.create(checkoutOptions);
     }
 
-    setCheckoutIdInStorage(checkoutResource.id);
+    setCheckoutIdInStorage(sessionToken, checkoutResource.id);
 
     setShop(shopInfo);
     setCheckout(checkoutResource);
@@ -101,8 +110,10 @@ export function CommerceProvider({ children, config }: CommerceProps) {
       client,
       checkout,
       shop,
-      currencyCode: config.currencyCode,
-      updateCheckout: updateCheckout
+      updateCheckout: updateCheckout,
+      currencyCode: config.currencyCode || checkout?.currencyCode,
+      locale,
+      sessionToken
     }),
     [client]
   );